fix(keycloak-theme): address all critical and major PR review findings (#179)
Critical:
- C1: Validate theme_primary_color against ^#[0-9a-fA-F]{6}$ before CSS
  interpolation in template.ftl; fallback to default brand green on mismatch
- C2: Add infra/keycloak/Dockerfile (COPY theme); build/push to GHCR in
  deploy-staging.yml; update deploy config to use custom image
- C3: Remove demo-asso from realm-givernance.json (shared across all envs);
  create it in keycloak-sync-realm.sh guarded by localhost-only check
- C4: Add Givernance shims for error.ftl, info.ftl, login-page-expired.ftl

Major:
- M1: Validate logo_url starts with https://; add referrerpolicy/crossorigin
- M2: Reject kc_org extraction in production unless hostname matches trusted
  domain suffixes (.givernance.org, .givernance.eu)
- M3: Replace Google Fonts @import with self-hosted @font-face (GDPR Art. 6)
- M6: Use color-mix(in srgb, <primary> 85%, black) for per-org hover color
- M7: Inject showPassword/hidePassword strings via window.gvI18n in template.ftl;
  update aria-labels and JS toggles in login.ftl / login-update-password.ftl
- M8: Restore locale picker in template.ftl for direct KC flows (password reset
  links etc.) where the app's /login page is bypassed
- M9: Hoist locale picker out of login/page.tsx; remove ?locale= URL param;
  rely on NEXT_LOCALE cookie set by setLocale() server action
- M10: Add auth.login.languageLabel to en.json / fr.json for LocalePicker
- M11: Extract LocalePicker to src/components/auth/locale-picker.tsx and hoist
  into (auth)/layout.tsx so all pre-auth pages share it

Minor:
- m1: Remove tabindex attributes from all login FTL forms
- m4: Remove duplicate doClickHere key from both message files
- m5: Replace display:none with off-screen positioning + aria-hidden on
  hidden username input in login-update-password.ftl
- m6: Make autocomplete conditional on realm.loginWithEmailAllowed in
  login-reset-password.ftl
- m7: Add staging/preview/dev/auth to NON_ORG set in extractOrgAlias
- m9: Remove empty logo_url attributes from realm-givernance.json
