+++ .github/workflows/deploy-staging.yml
@@ -62,6 +62,15 @@ jobs:
           KEYCLOAK_ADMIN_CLIENT_SECRET=$KC_CLIENT_SECRET
           EOF
 
+      - name: Build and push Keycloak image
+        run: |
+          echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u "${{ github.actor }}" --password-stdin
+          docker build \
+            -t ghcr.io/purposestack/givernance-keycloak:latest \
+            -f infra/keycloak/Dockerfile \
+            infra/keycloak/
+          docker push ghcr.io/purposestack/givernance-keycloak:latest
+
       - name: Release Deploy Lock
         run: kamal lock release -c config/deploy-staging.yml || true
