+++ infra/keycloak/themes/givernance/login/messages/messages_en.properties
@@ -17,6 +17,11 @@ doRegister=Create account
 doForgotPassword=Forgot password?
 doClickHere=click here
 doSubmit=Submit
+selectLocale=Language
+showPassword=Show password
+hidePassword=Hide password
+showPasswordConfirm=Show password confirmation
+hidePasswordConfirm=Hide password confirmation
 doBack=Back
 doCancel=Cancel
 doTryAgain=Try again
@@ -58,4 +63,3 @@ pageExpiredTitle=Session expired
 pageExpiredMsg1=To continue, please
 pageExpiredMsg2=or
 pageExpiredMsg3=restart the login process.
-doClickHere=click here
