+++ infra/keycloak/themes/givernance/login/messages/messages_fr.properties
@@ -17,6 +17,11 @@ doRegister=Cr\u00e9er un compte
 doForgotPassword=Mot de passe oubli\u00e9\u00a0?
 doClickHere=cliquez ici
 doSubmit=Valider
+selectLocale=Langue
+showPassword=Afficher le mot de passe
+hidePassword=Masquer le mot de passe
+showPasswordConfirm=Afficher la confirmation
+hidePasswordConfirm=Masquer la confirmation
 doBack=Retour
 doCancel=Annuler
 doTryAgain=R\u00e9essayer
@@ -65,4 +70,3 @@ pageExpiredTitle=Session expir\u00e9e
 pageExpiredMsg1=Pour continuer,
 pageExpiredMsg2=ou
 pageExpiredMsg3=recommencez le processus de connexion.
-doClickHere=cliquez ici
