+++ packages/web/messages/fr.json
@@ -26,6 +26,7 @@
       "securityHint": "Sécurisé via le fournisseur d'identité de votre organisation",
       "noAccount": "Pas encore de compte ? <link>Demander un accès</link>",
       "close": "Fermer",
+      "languageLabel": "Langue",
       "errors": {
         "authFailed": "Échec de l'authentification. Veuillez réessayer.",
         "generic": "Une erreur est survenue lors de la connexion. Veuillez réessayer.",
